Extracting Work from Model
We need linked_resources and people to look like a model, but they
are not works.
