+++ app/repository_models/curation_concern/with_related_works.rb
@@ -12,7 +12,7 @@ module CurationConcern
       accepts_nested_attributes_for :related_works, :allow_destroy => true
 
       def related_work_tokens=(tokens)
-        self.related_work_ids = CurationConcern::Model.ids_from_tokens(tokens)
+        self.related_work_ids = CurationConcern::Work.ids_from_tokens(tokens)
       end
 
       def related_work_tokens
