+++ app/repository_models/dataset.rb
@@ -1,5 +1,5 @@
 class Dataset < ActiveFedora::Base
-  include CurationConcern::Model
+  include CurationConcern::Work
   include CurationConcern::WithGenericFiles
   include CurationConcern::WithLinkedResources
   include CurationConcern::WithRelatedWorks
