+++ app/repository_models/generic_work.rb
@@ -1,5 +1,5 @@
 class GenericWork < ActiveFedora::Base
-  include CurationConcern::Model
+  include CurationConcern::Work
   include CurationConcern::WithGenericFiles
   include CurationConcern::WithLinkedResources
   include CurationConcern::WithLinkedContributors
